[#1] Setup Nested Component
