+++ src/App.css
@@ -2,6 +2,35 @@
   text-align: center;
 }
 
+.Card {
+  border: 10px solid lightgray;
+  padding: 20px;
+  margin-bottom: 40px;
+}
+
+.Card .Card {
+  border: none;
+  outline: 2px solid orange;
+}
+.Card .Card .Card {
+  outline: 2px solid red;
+}
+.Card .Card .Card .Card {
+  outline: 2px solid black;
+}
+
+.Card > .Card .children {
+  display: flex;
+  flex-direction: column;
+  padding: 10px;
+}
+
+.Card .children {
+  display: flex;
+  flex-direction: column;
+  padding: 10px;
+}
+
 .App-logo {
   height: 40vmin;
 }
